+++ tools/test-harnesses/README.md
@@ -1,3 +1,3 @@
-# `@lg-tools/test-harness`
+# `@lg-tools/test-harnesses`
 
 This package provides several fundamental queries that will be utilized by every test harness we ship.
